Move functions into api
