+++ functions/api/ping.ts
