+++ tests/test-env.d.ts
@@ -1,4 +1,4 @@
-import { Env } from '@/types'
+import { Env } from '@/common/types'
 
 declare module 'cloudflare:test' {
   interface ProvidedEnv extends Env {}
